2nd code review attempt.
CODING STANDARS: fixed
COMMENTS: The "NOTES" comments are for other people to understand the current issue of the code. So I think it's reasonable to leave it there.
LOGIC: it's just style. I keep it the same as it is.
ERROR HANDLING: fixed
CODE DECISION: in order to keep the other developers have the flexibility to deal with initial DB and check table size, I will let those two functions public. Also, it's helpful to have those public for testing purpose.